requirements.txt
streamlit
streamlit-pills
llama-index==0.9.7
llama-hub==0.0.44
# NOTE: this is due to a trivial dependency in the web tool, will refactor
langchain==0.0.305
pypdf
telethon
python-dotenv
